review: 별도의 hook으로 분리
